Revamped use of --tag-cnid for BBC downloads
The get_iplayer implementation of --tag-cnid relied on an iTunes hack
that no longer works.  The get_iplayer implementation has been changed
upstream to generate a 32-bit hash from programme PIDs to use as the
catalogue ID (cnID atom) values.  This should prevent HD-SD combinations
from being mismatched, but there is no guarantee that the generated values
won't conflict with values for media purchased from the iTunes Store.

Because this is an esoteric feature that can cause headaches when it
doesn't work as expected, it is now opt-in via a hidden preference:

BBC_TagCNID (bool) : If YES, --tag-cnid will be added to the get_iplayer
command line by GiA.

git-svn-id: https://get-iplayer-automator.googlecode.com/svn/trunk@445 98b4c264-8670-11de-a133-9d545b4d7674